Move version information to be cmd-specific, not library specific;
If some other project is depending on this, it cannot retrieve meaningful information as there is no build procedure.
