+++ .goreleaser.yaml
@@ -16,7 +16,7 @@ builds:
     mod_timestamp: '{{ .CommitTimestamp }}'
 
     ldflags:
-      - -s -w -X 'github.com/dertseha/goconsider/pkg/consider.scmTag={{.Version}}' -X 'github.com/dertseha/goconsider/pkg/consider.scmCommit={{.Commit}}'
+      - -s -w -X 'github.com/dertseha/goconsider/internal/version.scmTag={{.Version}}' -X 'github.com/dertseha/goconsider/internal/version.scmCommit={{.Commit}}'
 
 archives:
   - format: tar.gz
