+++ internal/version/doc.go
@@ -1,0 +1,5 @@
+// Package version contains the build-information that can be re-used for distributable binaries.
+//
+// As the version information requires setting via the build system, it must be internal as it cannot be
+// used by library-dependents.
+package version
