+++ internal/version/info.go
@@ -1,11 +1,11 @@
-package consider
+package version
 
 // The following "constants" need to be initialized at build-time, through ldflags.
 var scmTag string
 var scmCommit string
 
-// VersionInfo describes the version of the library.
-type VersionInfo struct {
+// Info describes the version of the library.
+type Info struct {
 	// CoreWithPreRelease is the combination of the three numeric identifiers and an optional pre-release suffix.
 	CoreWithPreRelease string
 	// Build is the auxiliary information for the version.
@@ -13,7 +13,7 @@ type VersionInfo struct {
 }
 
 // Version returns the version specified at build time.
-func Version() VersionInfo {
+func Version() Info {
 	coreWithPreRelease := scmTag
 	if len(coreWithPreRelease) == 0 {
 		coreWithPreRelease = "0.0.0-alpha"
@@ -22,7 +22,7 @@ func Version() VersionInfo {
 	if len(build) == 0 {
 		build = "manual"
 	}
-	return VersionInfo{
+	return Info{
 		CoreWithPreRelease: coreWithPreRelease,
 		Build:              build,
 	}
