+++ pkg/analyzer/analyzer.go
@@ -42,12 +42,10 @@ func NewAnalyzerFromFlags() *analysis.Analyzer {
 }
 
 func newBaseAnalyzer() *analysis.Analyzer {
-	an := &analysis.Analyzer{
+	return &analysis.Analyzer{
 		Name: analyzerName,
 		Doc:  documentation,
 	}
-	an.Flags.Var(versionFlag{}, "V", "print version and exit")
-	return an
 }
 
 func runnerWithSettingsFrom(factory func() (consider.Settings, error)) func(*analysis.Pass) (interface{}, error) {
